카드 뒤집기 애니메이션 추가
카드 뒤집기 애니메이션을 추가하였습니다.
추가로 스타트씬 디자인 적용하였습니다.
