+++ Assets/Animations/Card-Idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aa0e3614f31cb4c4e9c949aef3390cf3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
