+++ Assets/Animations/Card.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa6d20996a8389643a20c021451db26d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
