+++ Assets/Animations/Card_Flip.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8a24c308995951747acab41b2086b3ca
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
