+++ Assets/Animations/StartImage.anim
@@ -22,12 +22,14 @@ AnimationClip:
     curve:
     - time: 0
       value: {fileID: 21300000, guid: 07abac1c1584a644d911075a869488c6, type: 3}
-    - time: 0.21666667
+    - time: 0.25
       value: {fileID: 21300000, guid: dccbfdac83f2e834ab955dddd0730111, type: 3}
-    - time: 0.45
-      value: {fileID: 21300000, guid: 52bdbe45c4029a548a5f5d765d58e62b, type: 3}
-    - time: 0.6666667
+    - time: 0.5
+      value: {fileID: 21300000, guid: 5ab66f27fdd384349ad6b15e8cdc4da9, type: 3}
+    - time: 0.75
       value: {fileID: 21300000, guid: 0d5a7031d938de248bf8781bca59d587, type: 3}
+    - time: 1
+      value: {fileID: 21300000, guid: 07abac1c1584a644d911075a869488c6, type: 3}
     attribute: m_Sprite
     path: 
     classID: 114
@@ -52,14 +54,15 @@ AnimationClip:
     pptrCurveMapping:
     - {fileID: 21300000, guid: 07abac1c1584a644d911075a869488c6, type: 3}
     - {fileID: 21300000, guid: dccbfdac83f2e834ab955dddd0730111, type: 3}
-    - {fileID: 21300000, guid: 52bdbe45c4029a548a5f5d765d58e62b, type: 3}
+    - {fileID: 21300000, guid: 5ab66f27fdd384349ad6b15e8cdc4da9, type: 3}
     - {fileID: 21300000, guid: 0d5a7031d938de248bf8781bca59d587, type: 3}
+    - {fileID: 21300000, guid: 07abac1c1584a644d911075a869488c6, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.68333334
+    m_StopTime: 1.0166667
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
