+++ Assets/Resources/Font/BMHANNA_11yrs_ttf SDF_B.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: d7765b501a1a4f64ea2e0cb5c151d853
+guid: 05cacbff03e7015459e95d2938f55cf4
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
