+++ Assets/Resources/Images/TeamEight5 1.jpg.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5159dd10dea86134eb8ae35a8e31b103
+guid: 5ab66f27fdd384349ad6b15e8cdc4da9
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
