+++ Assets/Resources/Images/함정카드.jpg.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5159dd10dea86134eb8ae35a8e31b103
+guid: 887a32b630e635d49a841a2cec076deb
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
