--- Assets/Scripts/CardArray.cs
@@ -1,25 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Card : MonoBehaviour
-{
-    int idx = 0;
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    public void Setting(int number)
-    {
-      idx = number;
-        Resources.Load<Sprite>($"TeamEight{idx}");
-    }
-}
