+++ Assets/Scripts/UI_Base.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class UI_Base : MonoBehaviour
+{
+    void Start()
+    {
+        
+    }
+
+    public void OnApplicationQuit()
+    {
+        Application.Quit();
+    }
+
+}
