+++ Assets/Scripts/UI_Base.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30af942e1aaffed4a80976ad7fca7c04
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
